move: dispatcher di 관련 코드 core:android 모듈로 이동
