+++ core/android/src/test/java/com/susu/core/android/SusuDispatcher.kt
@@ -1,4 +1,4 @@
-package com.susu.core.common
+package com.susu.core.android
 
 import javax.inject.Qualifier
